tests/wpt/web-platform-tests/DOMEvents/tests/approved/ProcessingInstruction.DOMCharacterDataModified.html

<!DOCTYPE html>
<html>
<head>
<title> ProcessingInstruction.data and DOMCharacterDataModified event </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>

<iframe id="helper" style="display: none"></iframe>
        
<script>
    var description = "Test Description: " + 
                      "DOMCharacterDataModified event fires after ProcessingInstruction.data have been modified, " +
                      "but the node itself has not been inserted or deleted. The proximal event target of this " +
                      "event shall be the ProcessingInstruction node.";
    
    var t = async_test(description);
    
    var HELPER = document.getElementById("helper");
    
    HELPER.onload = t.step_func(function()
    {
        assert_true(HELPER.contentWindow.TestResult);
        t.done();
    });
    
    HELPER.src = "./support/ProcessingInstruction.DOMCharacterDataModified.xml";
</script>
</body>
</html>
